requirements.txt
py-cord
click
requests
pillow
